==> Codes/01 Data Prep - Mortality at county level by different causes.ipynb <==
# <span style="color:Maroon">Data Preparation -- Part 1

<span style="color:Green">__Overview:__ Create dataset at county level. The columns store the values of Mortality rate (mean, upper and lower bounds) for different causes.

# Import required libraries
import numpy as np
import pandas as pd
import os as os
import xlrd

os.chdir("..")
cwd = os.getcwd()

# Define the directories based on type of data
country_category_year = "Data\\country_category_year"
labels = "Data\\labels"
state_category_year = "Data\\state_category_year"
states_gender_year_cause = "Data\\states_gender_year_cause"

# define the filenames based on type of data
file_coun = "IHME_USA_COUNTY_MORTALITY_RATES_1980_2014_NATIONAL_Y2016M12D13.xlsx"
file_label = "IHME_USA_COUNTY_MORTALITY_RATES_1980_2014_CODEBOOK_Y2017M05D19.csv"
file_st_ct_yr1 = "IHME_USA_COUNTY_MORTALITY_RATES_1980_2014_"
file_st_ct_yr2 = "_Y2016M12D13"
file_st_gn_yr_cs1 = "IHME_USA_COUNTY_MORTALITY_RATES_1980_2014_"
file_st_gn_yr_cs2 = "_Y2017M05D19"

# List with state names
states = ["ALABAMA", "ALASKA", "ARIZONA", "ARKANSAS", "CALIFORNIA", "COLORADO", "CONNECTICUT", "DELAWARE", 
          "DISTRICT_OF_COLOMBIA", "FLORIDA", "GEORGIA", "HAWAII", "IDAHO", "ILLINOIS", "INDIANA", "IOWA", "KANSAS", 
          "KENTUCKY", "LOUISIANA", "MAINE", "MARYLAND", "MASSACHUSETTS", "MICHIGAN", "MINNESOTA", "MISSISSIPPI",
          "MISSOURI", "MONTANA", "NEBRASKA", "NEVADA", "NEW_HAMPSHIRE", "NEW_JERSEY", "NEW_MEXICO", "NEW_YORK",
          "NORTH_CAROLINA", "NORTH_DAKOTA", "OHIO", "OKLAHOMA", "OREGON", "PENNSYLVANIA", "RHODE_ISLAND", 
          "SOUTH_CAROLINA", "SOUTH_DAKOTA", "TENNESSEE", "TEXAS", "UTAH", "VERMONT", "VIRGINIA", "WASHINGTON", 
          "WEST_VIRGINIA", "WISCONSIN", "WYOMING"]

# List with different categories
categories = ["HIV-AIDS & tuberculosis", "Diarrhea, lower respiratory", "Neglected tropical diseases",
              "Maternal disorders", "Neonatal disorders", "Nutritional deficiencies", "Other communicable, maternal, ",
              "Neoplasms", "Cardiovascular diseases", "Chronic respiratory diseases", "Cirrhosis & other chronic ",
              "Digestive diseases", "Neurological disorders", "Mental & substance use ", "Diabetes, urogenital, blood",
              "Musculoskeletal disorders", "Other non-communicable ", "Transport injuries", "Unintentional injuries",
              "Self-harm & interpersonal ", "Forces of nature, war, & legal "]

# Dictionary mapping excel tab name to acronym
category_dict = {"HVT":"HIV-AIDS & tuberculosis",
                 "DLR":"Diarrhea, lower respiratory",
                 "NTD":"Neglected tropical diseases",
                 "MDA":"Maternal disorders",
                 "NDA":"Neonatal disorders",
                 "NEA":"Nutritional deficiencies",
                 "OTA":"Other communicable, maternal, ",
                 "NOA":"Neoplasms",
                 "CDA":"Cardiovascular diseases",
                 "CRE":"Chronic respiratory diseases",
                 "COC":"Cirrhosis & other chronic ",
                 "DDA":"Digestive diseases", 
                 "NIA":"Neurological disorders",
                 "MSU":"Mental & substance use ",
                 "DUB":"Diabetes, urogenital, blood",
                 "MIA":"Musculoskeletal disorders",
                 "ONC":"Other non-communicable ",
                 "TIA":"Transport injuries",
                 "UIA":"Unintentional injuries",
                 "SHI":"Self-harm & interpersonal ",
                 "FON":"Forces of nature, war, & legal "}

category_key = ["HVT","DLR","NTD","MDA","NDA","NEA","OTA","NOA","CDA","CRE","COC","DDA","NIA","MSU","DUB","MIA","ONC","TIA"
               ,"UIA","SHI","FON"]

# Check if the list is unique (no two similar names)
len(category_key) == len(list(dict.fromkeys(category_key)))

country = "UNITED_STATES"

## <span style="color:Maroon">Read the label file

# Label file
os.chdir(f'{cwd}\\{labels}')
labels = pd.read_csv(file_label)
labels.head()

## <span style="color:Maroon">Read data at state level with different causes

def SplitCol_toRows(df, column_index):
    columns = df.columns
    data = df.copy()
    for i in range(0,len(column_index)):
        column_name = columns[column_index[i]]
        var1 = column_name + "_VAL"
        var2 = column_name + "_LB"
        var3 = column_name + "_UB"
        data[[var1, "tmp"]] = data[column_name].str.split("(", expand=True)
        data[[var2, var3]] = data["tmp"].str.split(",", expand=True)
        data[var3] = data[var2].str.replace(")",'')
        data[var1] = data[var1].astype(float)
        data[var2] = data[var2].astype(float)
        data[var3] = data[var3].astype(float)
        data = data.drop(["tmp",column_name], axis=1)
    return data

def perct_change(df, column_name1,column_name2,key):
    df[key+"_PC1980_2010"] = (df[column_name1]-df[column_name2])/df[column_name2]
    return df

# Data at country level
os.chdir(f'{cwd}\\{country_category_year}')

combined_data = pd.DataFrame()
for i in range(0, len(category_key)):
    sheetname = category_dict[category_key[i]]
    data = pd.read_excel(file_coun, sheet_name=sheetname,header=1,skipfooter=2, engine='openpyxl')
    list1 = ["MR1980","MR1985","MR1990","MR1995","MR2000","MR2005","MR2010"]
    list2 = ["MR2014","PC1980_2014"]
    list1 = [category_key[i]+"_"+sub for sub in list1]
    data.columns =list(data.columns[0:2])+list1+list2
    data = data.drop(list2,axis=1)
    column_index = [2,3,4,5,6,7,8]
    data = SplitCol_toRows(data,column_index)
    data = perct_change(data, category_key[i]+"_"+"MR2010_VAL",category_key[i]+"_"+"MR2005_VAL", category_key[i])
    data = data.drop("Location",axis=1)
    data["FIPS"] = data["FIPS"].fillna(0).astype(int)
    data = data.set_index("FIPS")
    combined_data = combined_data.merge(data,how="outer",left_index=True,right_index=True)

combined_data.head()

combined_data.to_csv("..//Model data//Mortality_Category_County_5yr_Data.csv")
