==> Codes/03 Data Prep County level population data .ipynb <==
# <span style="color:Maroon">Data Preparation - Part 3

<span style="color:Green">__Overview:__ Create dataset at county level. The columns store the data for census population levels of the county

# Import Required libraries
import pandas as pd
import numpy as np
import os as os

os.chdir("..")
cwd = os.getcwd()
other_data = "\\Data\\other data"
os.chdir(f'{cwd}{other_data}')

data = pd.read_csv("co-est2013-alldata.csv",encoding = "ISO-8859-1")

data.head()

data["FIPS"] = np.where(data["COUNTY"]==0,data["STATE"],data["STATE"]*1000+data["COUNTY"])

data.head()
data = data.drop(["SUMLEV","REGION","DIVISION","STATE","COUNTY","CTYNAME","STNAME"],axis=1)
data = data.set_index("FIPS")
data.to_csv(f"{cwd}\\Data\\Model data\\County_population_data.csv")
