==> Codes/04 Data Prep 04.ipynb <==
# <span style="color:Maroon">Data Preparation - Part 4

<span style="color:Green">__Overview:__ Create dataset at county level by dividing the population by race type

# Import libraries
import pandas as pd
import numpy as np
import os as os

os.chdir("..")
cwd = os.getcwd()
other_data = "\\Data\\other data"
os.chdir(f'{cwd}{other_data}')

data = pd.read_csv("cc-est2013-alldata.csv",encoding = "ISO-8859-1")

data.shape

# Drop 2013 data, as 2013 is the lag year
data = data[data["YEAR"]<6]
data = data[data["YEAR"]!=2]
data = data[data["YEAR"]!=3]

states = list(np.unique(np.array(data["STATE"])))

varlist = ["SUMLEV", "STATE", "COUNTY", "CTYNAME", "YEAR", "STNAME"]

years = [1,4,5]

def transpose_derive(df,year,state,county):
    df = df.set_index("AGEGRP")
    df = df.transpose()
    df['PC_OF_TOTAL'] = df[0]/df[0].max()
    for i in range(1,19):
        df["PC"+str(i)+"_"+str(2010+year)] = df[i]/df[0]
    vars = list(df.columns)
    cols = vars[0:19]
    vars = vars[19:]
    cols = [(str(2010+year)+"_"+str(col)) for col in cols]
    vars = cols+vars
    df.columns = vars
    categories = list(dict.fromkeys(list(df.index)))
    out_data = pd.DataFrame()
    for i in range(0,len(categories)):
        df1 = df[df.index==categories[i]]
        cols = list(df1.columns)
        cols = [(categories[i]+str(col)) for col in cols]
        df1.columns = cols 
        df1.index = [state*1000+county]
        out_data =out_data.merge(df1, how="outer", left_index=True, right_index=True)
    return out_data

out_data = pd.DataFrame()
for i in range(0,len(states)):
    data1 = data[data["STATE"]==states[i]]
    county = list(np.unique(np.array(data1["COUNTY"])))
    for j in range(0,len(county)):
        data2 = data1[data1["COUNTY"]==county[j]]
        outp = pd.DataFrame()
        for k in range(0,len(years)):
            data3 = data2[data2["YEAR"]==years[k]]
            data3 = data3.drop(varlist,axis=1)
            tmp = transpose_derive(data3,k,states[i],county[j])
            outp = outp.merge(tmp, how="outer", left_index=True, right_index=True)
        out_data = out_data.append(outp)

out_data.shape

out_data.head()

out_data.to_csv(f"{cwd}\\Data\\Model data\\Population_Race.csv")
