==> Codes/05 Data Prep for Model.ipynb <==
# <span style="color:Maroon"> Consolidate data for Model

# Required libraries
import pandas as pd
import numpy as np
import os as os

os.chdir("..")
cwd = os.getcwd()
direc = "C:/Users/sidhu/Downloads/Prudential/data/Model data"
os.chdir(f'{cwd}\\Data\\Model data\\')

### Process County level Population

data1 = pd.read_csv("County_population_data.csv")

pd.DataFrame(data1.columns).to_csv("data1_columns.csv",index=False)

data1.head()

varlist = ["ESTIMATESBASE2010", "POPESTIMATE2010", "POPESTIMATE2013", "NPOPCHG_2010", "NPOPCHG_2011", "NPOPCHG_2012",
           "NPOPCHG_2013", "BIRTHS2013", "DEATHS2013", "NATURALINC2013", "INTERNATIONALMIG2013", "DOMESTICMIG2013",
           "NETMIG2013", "RESIDUAL2013", "GQESTIMATESBASE2010", "GQESTIMATES2010", "GQESTIMATES2011", "GQESTIMATES2012",
           "GQESTIMATES2013", "RBIRTH2013", "RDEATH2013", "RNATURALINC2013", "RINTERNATIONALMIG2013", "RDOMESTICMIG2013",
           "RNETMIG2013"]
data1 = data1.drop(varlist,axis=1)

data1 = data1.set_index("FIPS")

data1.head()

### Processs Mortality: Category by County 5yr Data

data2 = pd.read_csv("Mortality_Category_County_5yr_Data.csv")

pd.DataFrame(data2.columns).to_csv("data2_columns.csv",index=False)

data2 = data2.set_index("FIPS")
data2.head()

### Process Mortality: Yearly by Category by Gender

data3_0 = pd.read_csv("Mortality_yearly_category_gender_part0.csv")
data3_1 = pd.read_csv("Mortality_yearly_category_gender_part1.csv")
data3_2 = pd.read_csv("Mortality_yearly_category_gender_part2.csv")
data3_3 = pd.read_csv("Mortality_yearly_category_gender_part3.csv")
data3_4 = pd.read_csv("Mortality_yearly_category_gender_part4.csv")

data3 = data3_0.append(data3_1).append(data3_2).append(data3_3)
del data3_0, data3_1, data3_2, data3_3

data3.columns = ["FIPS"] + list(data3.columns)[1:]

pd.DataFrame(data3.columns).to_csv("data3_columns.csv",index=False)

data3 = data3.set_index("FIPS")
data3.head()

varlist = ["ACA_3_2013", "ACA_1_2013", "ACA_2_2013", "CMN_3_2013", "CMN_1_2013", "CMN_2_2013", "HVT_3_2013", "HVT_1_2013",
           "HVT_2_2013", "DLR_3_2013", "DLR_1_2013", "DLR_2_2013", "NTD_3_2013", "NTD_1_2013", "NTD_2_2013", "MDA_3_2013",
           "MDA_2_2013", "NDA_3_2013", "NDA_1_2013", "NDA_2_2013", "NEA_3_2013", "NEA_1_2013", "NEA_2_2013", "OTA_3_2013",
           "OTA_1_2013", "OTA_2_2013", "OCD_3_2013", "OCD_1_2013", "OCD_2_2013", "NOA_3_2013", "NOA_1_2013", "NOA_2_2013",
           "CDA_3_2013", "CDA_1_2013", "CDA_2_2013", "CRE_3_2013", "CRE_1_2013", "CRE_2_2013", "COC_3_2013", "COC_1_2013",
           "COC_2_2013", "DDA_3_2013", "DDA_1_2013", "DDA_2_2013", "NIA_3_2013", "NIA_1_2013", "NIA_2_2013", "MSU_3_2013",
           "MSU_1_2013", "MSU_2_2013", "DUB_3_2013", "DUB_1_2013", "DUB_2_2013", "MIA_3_2013", "MIA_1_2013", "MIA_2_2013",
           "ONC_3_2013", "ONC_1_2013", "ONC_2_2013", "IAA_3_2013", "IAA_1_2013", "IAA_2_2013", "TIA_3_2013", "TIA_1_2013",
           "TIA_2_2013", "UIA_3_2013", "UIA_1_2013", "UIA_2_2013", "SHI_3_2013", "SHI_1_2013", "SHI_2_2013", "FON_3_2013",
           "FON_1_2013", "FON_2_2013"]

data3 = data3.drop(varlist,axis=1)

data3["Mortality_change_2012_2014"] = (data3["ACA_3_2014"] - data3["ACA_3_2012"])/data3["ACA_3_2012"]
data3["Mortality_change_2012_2014"].describe()

data3["Target"] = np.where(data3["Mortality_change_2012_2014"]>0.015,1,0)
data3["Target"].describe()

varlist = ["ACA_3_2014", "ACA_1_2014", "ACA_2_2014", "CMN_3_2014", "CMN_1_2014", "CMN_2_2014", "HVT_3_2014", "HVT_1_2014", 
           "HVT_2_2014", "DLR_3_2014", "DLR_1_2014", "DLR_2_2014", "NTD_3_2014", "NTD_1_2014", "NTD_2_2014", "MDA_3_2014",
           "MDA_2_2014", "NDA_3_2014", "NDA_1_2014", "NDA_2_2014", "NEA_3_2014", "NEA_1_2014", "NEA_2_2014", "OTA_3_2014",
           "OTA_1_2014", "OTA_2_2014", "OCD_3_2014", "OCD_1_2014", "OCD_2_2014", "NOA_3_2014", "NOA_1_2014", "NOA_2_2014",
           "CDA_3_2014", "CDA_1_2014", "CDA_2_2014", "CRE_3_2014", "CRE_1_2014", "CRE_2_2014", "COC_3_2014", "COC_1_2014",
           "COC_2_2014", "DDA_3_2014", "DDA_1_2014", "DDA_2_2014", "NIA_3_2014", "NIA_1_2014", "NIA_2_2014", "MSU_3_2014",
           "MSU_1_2014", "MSU_2_2014", "DUB_3_2014", "DUB_1_2014", "DUB_2_2014", "MIA_3_2014", "MIA_1_2014", "MIA_2_2014",
           "ONC_3_2014", "ONC_1_2014", "ONC_2_2014", "IAA_3_2014", "IAA_1_2014", "IAA_2_2014", "TIA_3_2014", "TIA_1_2014",
           "TIA_2_2014", "UIA_3_2014", "UIA_1_2014", "UIA_2_2014", "SHI_3_2014", "SHI_1_2014", "SHI_2_2014", "FON_3_2014",
           "FON_1_2014", "FON_2_2014"]

data3 = data3.drop(varlist, axis=1)

pd.DataFrame(data3.columns).to_csv("data3_columns.csv", index=False)

### Population by Race by Year

data4 = pd.read_csv("Population_Race.csv")

data4.shape

data4.head()

cols = list(data4.columns)[1:]
cols = ["FIPS"] + cols
data4.columns = cols
data4 = data4.set_index("FIPS")

data4.head()

pd.DataFrame(data4.columns).to_csv("Data4_columns.csv",index=False)

### Merge Datasets

data = data1.merge(data2, how="outer", left_index = True, right_index = True)

data = data.merge(data3, how="outer", left_index = True, right_index = True)

data = data.merge(data4, how="outer", left_index = True, right_index = True)

data.shape

data.head()

data = data[data.index > 100]

data = data.fillna(0)

data.to_csv("Final_model_dataset.csv")
